BuyAm/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
argon2-cffi==23.1.0
python-dotenv==1.0.0
itsdangerous==2.1.2
wtforms~=3.1.2
werkzeug~=3.0.6
alembic~=1.14.1
SQLAlchemy~=2.0.37